backend/requirements.txt
flask
flask_cors
torch
transformers
Pillow
numpy
requests
sentence-transformers